fix: Move vite to devDependencies for sdk package (#28)
We don't need vite as a runtime dependency in the sdk, and the
version is different for the operator repo. This is causing the
insall step to fail with

`npm error Cannot set properties of null (setting 'dev')`

when linking the web-applet/sdk code to a local operator package.

Move vite to devDependencies so it doesn't try install when in
operator.