fix(inspector): Add validation and formatting to params textarea
When a user submits invalid json, we don't tell them and instead
submit the bad data.

Add a change event to the textarea (a change event on a textarea
won't trigger for every character change, only blurring) that
tries to parse and format the json. If it fails, set a custom
validity message on the textarea (to block form submission) and
render a custom error message under the field with the json
parsing error.