feat(inspector): Create a model file for interacting with ai model
This is a replica of _some_ of the model setup from operator.

Given an applet we build:

* the system prompt, noting the actions the applet has
* a json response schema, for selecting the action and params

We then send those with our model and prompt and return the
result.

**Note**: We are re-creating the openAI provider and the model
each time, this could be improved, but I wanted to make sure we're
always in sync with the currently set token.

We might also want to sync with the applet from this file, but
currently that is the job of the consumer.