fix(inspector): Remove height and allow container to overflow.
We want the main container to grow to fill the rest of the screen,
not 100% height, as we also have the header taking up space.

We also want to enable overflow on app-viewer and the data view,
this way we keep our "footers" in place but allow the inner
windows to scroll.